fix: init translate service
