+++ src/app/core/components/header/header.component.scss
@@ -34,7 +34,7 @@
   }
 
   app-language {
-    margin-right: 30px;
+    margin-right: 40px;
   }
 
   @media (max-width: 820px) {
