+++ src/app/core/components/header/language/language.component.scss
@@ -5,6 +5,7 @@ form {
 label {
     text-decoration: none;
     font-size: 20px;
+    margin-left: 20px;
     cursor: pointer;
 }
 
@@ -26,6 +27,7 @@ mat-select {
 @media (max-width: 820px) {
     label {
         font-size: 15px;
+        margin-left: 10px;
     }
 
     mat-form-field {
