+++ src/app/core/components/header/language/language.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { FormGroup, FormBuilder } from '@angular/forms';
-import { TranslateService } from '@ngx-translate/core';
+import { OurTranslateService } from '../../../../shared/services/translate.service';
 
 @Component({
   selector: 'app-language',
@@ -9,24 +9,18 @@ import { TranslateService } from '@ngx-translate/core';
 })
 export class LanguageComponent {
   
-  private defaultLang = 'en';
-  public currentLang: string = localStorage.getItem('lang') || this.defaultLang;
   public languageForm: FormGroup;
+  public currentLang: string;
 
-  constructor(private translateService: TranslateService, private fb: FormBuilder) {
+  constructor(private translator: OurTranslateService, private fb: FormBuilder) {
     this.initForm();
-    this.setLanguage(this.currentLang);
-    this.languageForm.get('currentLanguage').valueChanges.subscribe(language => this.setLanguage(language));
+    this.currentLang = this.translator.currentLang;
+    this.languageForm.get('currentLanguage').valueChanges.subscribe(language => this.translator.setLanguage(language));
   }
 
   public initForm() {
     this.languageForm = this.fb.group({
-      currentLanguage: this.currentLang
+      currentLanguage: this.translator.currentLang
     });
-  }
-
-  public setLanguage(language: string) {
-    this.translateService.setDefaultLang(language);
-    localStorage.setItem('lang', language);
   }
 }
